second last changes

+++ models/dbHelper/helper.js
@@ -1,19 +1,20 @@
 const bcrypt = require("bcryptjs");
 const { UserDetails, UserRole, UserAccount } = require("../index");
+const { SendResponse } = require("../../utils/utils");
 
 const FindRoleByName = async (roleName) => {
-  return UserRole.findOne({
-    where: { Role: roleName },
-  });
+  return UserRole.findOne({ where: { Role: roleName }, });
 };
 
 const GenerateHashPassword = async (password) => {
+
   const salt = await bcrypt.genSalt(10);
 
   return bcrypt.hash(password, salt);
 };
 
 const SignUpUserDetails = async (userId, email, password, firstName, lastName) => {
+
   const hashedPassword = await GenerateHashPassword(password);
 
   return UserDetails.create({
@@ -26,43 +27,22 @@ const SignUpUserDetails = async (userId, email, password, firstName, lastName) =
 };
 
 const SignUpUserAccount = async (roleId) => {
-  return UserAccount.create({
-    RoleId: roleId,
-  });
+  return UserAccount.create({ RoleId: roleId, });
 };
 
 const FindUserById = async (UserId) => {
-  return UserAccount.findOne({ where: { UserId } })
-}
+  return UserAccount.findOne({ where: { UserId } });
+};
 
 const FindUserByEmail = async (Email) => {
-  return UserDetails.findOne({ where: { Email } })
-}
+  return UserDetails.findOne({ where: { Email } });
+};
 
 const FindUser = async (userDetails) => {
   return UserAccount.findByPk(userDetails.UserId);
-}
+};
 
 const FetchUserDetails = async (UserId) => {
-  // const response = await UserRole.findAll({});
-  // const user = await UserDetails.findOne({
-  //   where: { UserId },
-  //   include: [{
-  //     model: UserAccount,
-  //     include: [{
-  //       model: UserRole,
-  //       attributes: ['Role']
-  //     }]
-  //   }]
-  // });
-
-  // const response = await UserAccount.findByPk(UserId)
-  // const RoleId = response.RoleId;
-  // console.log(response);
-  // const role = await UserRole.findOne({
-  //   where: { RoleId }
-  // })
-
 
   const user = await UserDetails.findOne({
     where: { UserId },
@@ -78,13 +58,23 @@ const FetchUserDetails = async (UserId) => {
     throw new Error("User Not Found!");
   }
 
+  // check if user account is disabled or deleted then give me a message....
+  const IsDeleted = user.UserAccount.IsDeleted;
+  const IsDisabled = user.UserAccount.IsDisabled;
+
+  const message = IsDeleted ? "User Account is Deleted" : IsDisabled ? "User Account is Disabled" : null;
+
+  if (message) {
+    throw new Error(message);
+  }
+
   return {
     Email: user.Email,
     FirstName: user.FirstName,
     LastName: user.LastName,
     RoleName: user.UserAccount?.UserRole.Role ?? "null"
   };
-}
+};
 
 const FetchUserByIdAndDelete = async (UserId) => {
   const user = await UserAccount.findOne({ where: { UserId } });
@@ -95,7 +85,7 @@ const FetchUserByIdAndDelete = async (UserId) => {
   await user.update({ IsDeleted: true });
   // console.log("after", user);
   return;
-}
+};
 
 const UpdateUserStatus = async (UserId, status) => {
 
@@ -112,9 +102,15 @@ const UpdateUserStatus = async (UserId, status) => {
 const FindUserByParam = async (UserId) => {
   const user = await UserDetails.findOne({ where: { UserId } });
   return user;
-}
+};
 
 const DeleteUserDetails = async (UserId) => {
+
+  const user = await UserDetails.findOne({ where: { UserId } });
+
+  if (!user) {
+    throw new Error("User Not Found!")
+  }
 
   await UserDetails.destroy({
     where: {
@@ -122,27 +118,12 @@ const DeleteUserDetails = async (UserId) => {
     }
   });
 
-  // const userAccount = await UserAccount.findOne({
-  //   where: { UserId },
-  //   include: [{
-  //     model: UserRole,
-  //     attributes: ['Role'],
-  //   }]
-  // });
-  // const Role = userAccount?.UserRole?.Role ?? "not found";
-
-  // await UserRole.destroy({
-  //   where: {
-  //     Role: Role
-  //   }
-  // })
-
   await UserAccount.destroy({
     where: { UserId }
   });
 
   return;
-}
+};
 
 module.exports = {
   GenerateHashPassword,
